bee-init: some improvements and default template
added option --force .. use for beefile creation even if beefile exists

added hardcoded default template

+++ src/beeinit.sh
@@ -11,6 +11,7 @@ usage() {
     echo "options.."
     echo "  -h                .. show this help"
     echo "  -t [xorg|default] .. specify template used for beefile"
+    echo "  -f                .. force overwriting of beefile"
 }
 
 initialize() {
@@ -26,18 +27,28 @@ initialize() {
         fi
     fi
 
+    if [ -e ${pname}.bee ] && [ "$OPT_FORCE" != "yes" ] ; then
+        echo "${pname}.bee already exists .. use option -f to overwrite .."
+        exit 1
+    fi
+
     echo "creating ${pname}.bee with TEMPLATE='${TEMPLATE}' and SRCURL='${surl}'"
 
-    cp ${TEMPLATEPATH}${TEMPLATE} ${pname}.bee
-
-    sed -e "s,SRCURL.*,SRCURL[0]=\"${surl}\"," -i ${pname}.bee
-
+    if [ -e ${TEMPLATEPATH}${TEMPLATE} ] ; then
+        cp ${TEMPLATEPATH}${TEMPLATE} ${pname}.bee
+        sed -e "s,SRCURL.*,SRCURL[0]=\"${surl}\"," -i ${pname}.bee
+    else
+        cat >${pname}.bee<<"EOT"
+#!/bin/env beesh
+SRCURL[0]="${surl}"
+EOT
+    fi
     chmod 755 ${pname}.bee
 }
 
 options=$(getopt -n beeinit \
-                 -o ht: \
-                 --long help,template \
+                 -o ht:f \
+                 --long help,template,force \
                  -- "$@")
 if [ $? != 0 ] ; then
   usage
@@ -46,29 +57,33 @@ fi
 eval set -- "${options}"
 
 while true ; do
-  case "$1" in
-    -h|--help)
-      usage
-      exit 0
-    ;;
-    -t|--template)
-      shift
-      TEMPLATE=$1
-      if [ ! -e ${TEMPLATEAPTH}${TEMPLATE} ] ; then
-          echo "template '${TEMPLATE}' doesn't exist in '${TEMPLATEPATH}' .. 'default' template used instead .."
-          TEMPLATE=default
-      fi
-      shift
-      ;;
-    *)
-      shift
-      if [ -z ${1} ] ; then
-           usage
-           exit 1
-      fi
-      : ${TEMPLATE:=default}
-      initialize ${@}
-      exit 0;
-      ;;
-  esac
+    case "$1" in
+        -h|--help)
+            usage
+            exit 0
+            ;;
+        -t|--template)
+            shift
+            TEMPLATE=$1
+            if [ ! -e ${TEMPLATEAPTH}${TEMPLATE} ] ; then
+                echo "ignoring non-existant template '${TEMPLATEPATH}${TEMPLATE}' .."
+                unset TEMPLATE
+            fi
+            shift
+            ;;
+        -f|--force)
+            shift
+            OPT_FORCE="yes"
+            ;;
+        *)
+            shift
+            if [ -z ${1} ] ; then
+                 usage
+                 exit 1
+            fi
+            : ${TEMPLATE:=default}
+            initialize ${@}
+            exit 0;
+            ;;
+    esac
 done
